% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_offsets.R
\name{calculate_offsets}
\alias{calculate_offsets}
\title{Calculate Offsets}
\usage{
calculate_offsets(fit, rarray = rarray, tarray = tarray, X1 = NULL, X2 = NULL)
}
\arguments{
\item{fit}{}

\item{rarray}{}

\item{tarray}{}

\item{X1}{}

\item{X2}{}
}
\description{
Calculate Offsets
}
