% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_multi_fit_joint_slow.R
\name{cmulti_fit_joint_slow}
\alias{cmulti_fit_joint_slow}
\title{cmulti_fit_joint_slow}
\usage{
cmulti_fit_joint_slow(
  Yarray,
  rarray,
  tarray,
  X1 = NULL,
  X2 = NULL,
  maxdistint = 10,
  tau_inits = NULL,
  phi_inits = NULL,
  method = "Nelder-Mead",
  ...
)
}
\arguments{
\item{Yarray}{}

\item{rarray}{}

\item{tarray}{}

\item{X1}{}

\item{X2}{}

\item{maxdistint}{}

\item{tau_inits}{}

\item{phi_inits}{}

\item{method}{}

\item{...}{}
}
\description{
cmulti_fit_joint_slow
}
