% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_point_counts.R
\name{gen_survey_data_frame}
\alias{gen_survey_data_frame}
\title{gen_survey_data_frame}
\usage{
gen_survey_data_frame(
  sim_rep,
  tau,
  phi,
  Density,
  distance_protocols,
  time_protocols
)
}
\arguments{
\item{sim_rep}{}

\item{tau}{}

\item{phi}{}

\item{Density}{}

\item{distance_protocols}{}

\item{time_protocols}{}
}
\description{
gen_survey_data_frame
}
