% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logdmultinomCPP}
\alias{logdmultinomCPP}
\title{logdmultinomCPP}
\usage{
logdmultinomCPP(x, size, prob)
}
\arguments{
\item{x}{A matrix}

\item{size}{A double}

\item{prob}{A Matrix}
}
\description{
A function to calculate log density
of multinomial.
}
