% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nll_fun}
\alias{nll_fun}
\title{Negative log likelihood function}
\usage{
nll_fun(
  params,
  X1,
  X2,
  tau_params,
  nsurvey,
  Yarray,
  tarray,
  rarray,
  nrint,
  ntint,
  max_r,
  Ysum,
  nlimit
)
}
\description{
A function to calculate negative log likelihood.
}
