% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_point_counts.R
\name{simulate_point_counts}
\alias{simulate_point_counts}
\title{simulate_point_counts}
\usage{
simulate_point_counts(survey_data_frame, ncores = 1)
}
\arguments{
\item{survey_data_frame}{}

\item{ncores}{}
}
\description{
simulate_point_counts
}
