% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_point_counts.R
\name{simulate_single_point_count}
\alias{simulate_single_point_count}
\title{simulate_single_point_count}
\usage{
simulate_single_point_count(
  tau,
  phi,
  Density,
  distance_protocols,
  time_protocols,
  seed,
  mdbin,
  mtbin,
  dim = 10,
  ...
)
}
\arguments{
\item{tau}{}

\item{phi}{}

\item{Density}{}

\item{distance_protocols}{}

\item{time_protocols}{}

\item{seed}{}

\item{mdbin}{}

\item{mtbin}{}

\item{dim}{}

\item{...}{}
}
\description{
simulate_single_point_count
}
