\name{cmulti_joint_dev-package}
\alias{cmulti_joint_dev-package}
\alias{cmulti_joint_dev}
\docType{package}
\title{
\packageTitle{cmulti_joint_dev}
}
\description{
\packageDescription{cmulti_joint_dev}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cmulti_joint_dev}
\packageIndices{cmulti_joint_dev}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{cmulti_joint_dev}

Maintainer: \packageMaintainer{cmulti_joint_dev}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
