\name{IlesShowUSomething-package}
\alias{IlesShowUSomething-package}
\alias{IlesShowUSomething}
\docType{package}
\title{
\packageTitle{IlesShowUSomething}
}
\description{
\packageDescription{IlesShowUSomething}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IlesShowUSomething}
\packageIndices{IlesShowUSomething}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{IlesShowUSomething}

Maintainer: \packageMaintainer{IlesShowUSomething}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, ~~
~~ from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
